Welcoming  our new maintainer and a few mix/ci-cd updates
we also included elixir 1.11 in the ci/cd matrix, because we recommend it as the minimum version in the mix. We also corrected the license name to get rid of the error message given by hex.pm, that of: license not recognized. 

And we have new maintainers, w⦿‿⦿t!! 

Greg, thanks for joining us, welcome!